+++ my_profile/src/app/page.tsx
@@ -1,5 +1,5 @@
 import { TopContent } from "../components/TopContent";
-import { Portfolio } from "../components/Portfolio";
+import { PortfolioAndBlog } from "../components/Portfolio";
 import { SkillsAndQualification } from "../components/SkillsAndQualification";
 import { Career } from "../components/Career";
 import { Working } from "../components/Working";
@@ -13,7 +13,7 @@ export default async function Home() {
   return (
     <main className="mt-[50px]">
       <TopContent topData={topData} />
-      <Portfolio />
+      <PortfolioAndBlog />
       <SkillsAndQualification />
       <Career />
       <Working />
